% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{add_to_list}
\alias{add_to_list}
\title{Add network to list}
\usage{
add_to_list(l, name, object)
}
\arguments{
\item{l}{list of networks}

\item{name}{character, name of newly created entry}

\item{object}{matrix containing the network to be added}
}
\value{
list containing the existing networks and the added network
}
\description{
This helper function used in the function
\code{create_statistical_networks_list} adds a network to a list of networks.
}
\details{
Used internally in \code{create_statistical_networks_list}
}
\examples{
add_to_list(l, name, object) 
data("x_test", package = "MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
cor_pearson <- correlation(x, type = "pearson")
cor_spearman <- correlation(x, type = "spearman")
l <- list(pearson = cor_pearson)
add_to_list(l, "spearman", cor_spearman)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
