% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{addToList}
\alias{addToList}
\title{Add adjacency matrix to list}
\usage{
addToList(l, name, object)
}
\arguments{
\item{l}{list of adjacency matrices}

\item{name}{character, name of newly created entry}

\item{object}{matrix containing the adjacency matrix to be added}
}
\value{
list containing the existing adjacency matrices and the added 
adjacency matrix
}
\description{
This helper function used in the function
\code{createStatisticalAdjacencyList} adds a adjacency matrix to a list of 
adjacency matrices.
}
\details{
Used internally in \code{createStatisticalAdjacencyList}
}
\examples{
data("x_test", package="MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
cor_pearson <- correlation(x, type="pearson")
cor_spearman <- correlation(x, type="spearman")
l <- list(pearson=cor_pearson)
addToList(l, "spearman", cor_spearman)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
