% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{adjacency_list}
\alias{adjacency_list}
\title{Create a list from adjacency matrices}
\usage{
adjacency_list(x, from)
}
\arguments{
\item{x}{`data.frame`, adjacency matrix, that has previously been generated 
by the functions `structural`, `statistical`, or `combine`}

\item{from}{`character` defines the function from which the adjacency matrix originated.
This can be either "structural", "statistical", or "combine".}
}
\value{
`data.frame` containing the listed values each pair of features from 
adjacency matrix `x` specified by `from`.
}
\description{
The function `adjacency_list` creates a dataframe from the adjacency matrix.
Unique values are displayed by the dataframe. Missing values and the upper triangle of 
the adjacency matrix are left out.
The function returns a dataframe containing different column outputs depending 
on the source of the adjacency matrix, which is defined by `from`.
}
\details{
exportNet2gml is a function that exports adjacency matrices to gml using igraph
Needs following attributes:
x: adjacency matrix that needs to be exported
from: originated from wich function, possible values are 
- "structural" Produces a gml file with edge attributes containing mass difference values, data saved as "structural_type.gml"
- "statistical+p" produces a gml file with edge attributes containing correlation values and p-values, saved as "statistical.'model'.gml"
(TO DO: ADD "statistical")
- "combine" produces a gml file with edge attributes containing correlation and p-values for pearson/ spearman correlations, 
saved as "combined.gml"


The function `adjacency_list` includes functionality to generate a dataframe from
adjacency matrices. The output depends on the original function from which the adjacency
matrix was generated.
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
stat_adj <- statistical(x = x, model = c("pearson", "spearman"))
adjacency_list(x = stat_adj, from = "statistical")

}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
