% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{annotaionNames}
\alias{annotaionNames}
\title{Add annotations to adjacency list}
\usage{
annotaionNames(list, names)
}
\arguments{
\item{`list`}{is the adjacency list, and `names` is a dataframe containing the feature names as rows,
mz values, RT values and annotations in columns}
}
\value{
`data.frame` containing the the `list`-values and additional column with annotations
}
\description{
The function `annotationNames` adds available annotation values for all features
to the adjacency list.
}
\details{
annotaionNames adds annotation to an adjacency list as additional column.
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
stat_adj <- statistical(x = x, model = c("pearson", "spearman"))
adj_l <- adjacency_list(x = stat_adj, from = "statistical")
annotationNames(list = adj_l, names = annotationFile)

}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
