% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{bayes}
\alias{bayes}
\title{Create a statistical network based on constraint-based structure 
learning algorithm}
\usage{
bayes(x, ...)
}
\arguments{
\item{x}{matrix, where columns are the samples and the rows are features 
(metabolites), cell entries are intensity values}

\item{...}{parameters passed to \code{fast.iamb}}
}
\value{
matrix, matrix with edges inferred from constraint-based structure 
learning algorithm \code{fast.iamb}
}
\description{
\code{bayes} infers a statistical network using 
constraint-based structure learning algorithm \code{fast.iamb} from the 
\code{bnlearn} package. \code{bayes} extracts then the reported 
connections from running the \code{fast.iamb} function and assigns the
arcs of the discrete Bayesian network to binary values into a matrix that is 
returned by \code{bayes}.
}
\details{
For use of the parameters used in the \code{fast.iamb} function, 
refer to ?bnlearn::fast.iamb.
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
bayes(x)
}
\author{
Thomas Naake, \email{thomasnaake @googlemail.com}
}
