% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_structural_statistical}
\alias{combine_structural_statistical}
\title{Combine structural and statistical network}
\usage{
combine_structural_statistical(structure, statistical, treshold = 1)
}
\arguments{
\item{structure}{matrix containing structural network}

\item{statistical}{matrix containing statistical netowrk}
}
\value{
a matrix containing the consensus network as described above 
harbouring connections reported by the structual network and statistcal 
networks.
}
\description{
The function \code{combine_structural_statistical} takes as 
input the structural and statistical network, that were create in former 
steps, adds them together and will report a connection between metabolites
when the edge is reported in networks. \code{combine_structural_statistical}
returns this consensus matrix of the structural and statistical network.
}
\details{
\code{combine_structural_statistical} will call internally the 
function \code{consensus_network} to create the consensus network of the 
statistical networks.
}
\examples{
#' data("x_test", package = "MetNet")
x_test <- as.matrix(x_test)
functional_groups <- rbind(
    c("Hydroxylation (–H)", "O", "15.9949146221"),
    c("Malonyl group (–H2O)", "C3H2O3", "86.0003939305"),
    c("C6H10O6", "C6H10O6", "178.0477380536"),
    c("D-ribose (–H2O) (ribosylation)", "C5H8O4", "132.0422587452"),
    c("Disaccharide (–H2O)", "C12H20O11", "340.1005614851"),
    c("Glucuronic acid (–H2O)", "C6H8O6", "176.0320879894"),
    c("Monosaccharide (–H2O)", "C6H10O5", "162.0528234315"),
    c("Trisaccharide (–H2O)", "C18H30O15", "486.1584702945"))
functional_groups <- data.frame(group = functional_groups[,1],
                                formula = functional_groups[,2],
                                mass = as.numeric(functional_groups[,3]))
struct_net <- create_structural_network(x_test, functional_groups, ppm = 5)
stat_net <- create_statistical_network(x_test, model = c("pearson", "spearman","bayes"), 
    adjust_correlation = "bonferroni")
combine_structural_statistical(struct_net[[1]], stat_net)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
