% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{consensus_network}
\alias{consensus_network}
\title{Create a consensus network of statistical networks}
\usage{
consensus_network(l, threshold = 1, ...)
}
\arguments{
\item{l}{list, each entry of the list contains a network}

\item{threshold}{numeric, when combining the networks the threshold 
parameter defines if an edge is reported or not. For 
\code{method = "central.graph"} threshold is set to 1 by default. For other
values of method, the value should be carefully defined by the user. If 
threshold is set to NULL (default), it will be set to 1 internally.}

\item{...}{parameters passed to the function \code{consensus} in the
\code{sna} package}
}
\value{
matrix, consensus matrix from statistical networks
}
\description{
The function takes a list of parameters (\code{l}) as input and
creates a consensus network from these networks by calling the function 
\code{consensus} from the \code{sna} package. Depending on the chosen 
\code{method } in \code{consensus}, the threshold of the consensus network
should be chosen accordingly to report a connection by different 
statistical methods.
}
\details{
\code{consensus_network} is a wrapper function of the 
\code{consensus} function of the \code{sna} package. For use of the 
parameters used in the \code{consensus} function, refer to ?sna::consensus.
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
stat_net_l <- create_statistical_networks_list(x, c("pearson", "spearman"))
consensus_network(stat_net_l)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
