% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{consensusAdjacency}
\alias{consensusAdjacency}
\title{Create a consensus adjacency matrix of statistical adjacency 
matrices}
\usage{
consensusAdjacency(l, threshold=1, ...)
}
\arguments{
\item{l}{list, each entry of the list contains an adjacency matrix}

\item{threshold}{numeric, when combining the adjacency matrices the 
\code{threshold} parameter defines if an edge is reported or not. For 
\code{method="central.graph"} threshold is set to 1 by default. For other
values of method, the value should be carefully defined by the user. If 
threshold is set to \code{NULL} (default), it will be set to 1 internally.}

\item{...}{parameters passed to the function \code{consensus} in the
\code{sna} package}
}
\value{
matrix, consensus matrix from adjacency matrices
}
\description{
The function takes a list of parameters (\code{l}) as input and
creates a consensus adjacency matrix from these adjacency matrices by 
calling the function \code{consensus} from the \code{sna} package. Depending 
on the chosen \code{method} in \code{consensus}, the threshold of the 
consensus adjacency matrix should be chosen accordingly to report a 
connection by different statistical methods.
}
\details{
\code{consensusAdjacency} is a wrapper function of the 
\code{consensus} function of the \code{sna} package. For use of the 
parameters used in the \code{consensus} function, refer to ?sna::consensus.
}
\examples{
data("x_test", package="MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
stat_adj_l <- createStatisticalAdjacencyList(x, c("pearson", "spearman"))
consensusAdjacency(stat_adj_l)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
