% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{correlation_p}
\alias{correlation_p}
\title{correlation_p is an additional function (complementary to correlation()), and needed for positive/negative pearson/spearman
correlation value and p-value calculation}
\usage{
correlation_p(x, type = "pearson", use = "pairwise.complete.obs")
}
\description{
correlation_p is an additional function (complementary to correlation()), and needed for positive/negative pearson/spearman
correlation value and p-value calculation
}
\details{
dependencies: Hmisc package
}
