% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{create_statistical_network}
\alias{create_statistical_network}
\title{Create statistical network}
\usage{
create_statistical_network(x, model, threshold = 1, ...)
}
\arguments{
\item{x}{matrix that contains intensity values of features/metabolites (rows)
per sample (columns).}

\item{model, }{character vector containing the model that will be used 
("lasso", "randomForest", "clr", "aracne", "pearson", "spearman", "bayes")}

\item{...}{parameters passed to the functions  \code{lasso}, 
\code{randomForest}, \code{clr}, \code{aracne}, \code{correlation},
\code{bayes} and/or \code{consensus_network}}
}
\value{
matrix, containing binary values if a connection is present or not
}
\description{
\code{create_statistical_network} creates a consensus network
given the models to use.
}
\details{
\code{create_statistical_network} is a wrapper function for the 
functions \code{create_statistical_networks_list} and 
\code{consensus_network}. See \code{?create_statistical_networks_list} and 
\code{?consensus_network} for further details. The function 
\code{create_statistical_networks_list}
includes functionality to caluclate statistical networks based on 
LASSO (L1 norm)-regression, random forests, context likelihood of 
relatedness (CLR), the algorithm for the reconstruction of accurate 
Spearman correlation and Constraint-based structure learning (Bayes).
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
create_statistical_network(x, c("pearson", "spearman"))
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
