% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{createStatisticalAdjacency}
\alias{createStatisticalAdjacency}
\title{Create statistical adjacency matrix}
\usage{
createStatisticalAdjacency(x, model, threshold = 1, ...)
}
\arguments{
\item{x}{matrix that contains intensity values of features/metabolites 
(rows) per sample (columns).}

\item{model, }{character, vector containing the model that will be used 
("lasso", "randomForest", "clr", "aracne", "pearson", "pearson_partial", 
"pearson_semipartial","spearman", "spearman_partial", 
"spearman_semipartial", "bayes")}

\item{threshold}{numeric, when combining the adjacency matrices the 
threshold parameter defines if an edge is reported or not. For 
\code{method = "central.graph"} threshold is set to 1 by default. For other
values of method, the value should be carefully defined by the user. If 
threshold is set to NULL (default), it will be set to 1 internally.}

\item{...}{parameters passed to the functions  \code{lasso}, 
\code{randomForest}, \code{clr}, \code{aracne}, \code{correlation},
\code{bayes} and/or \code{consensusAdjacency}}
}
\value{
matrix, containing binary values if a connection is present or not
}
\description{
\code{createStatisticalAdjacency} creates a consensus 
adjacency matrix given the models to use.
}
\details{
\code{createStatisticalAdjacency} is a wrapper function for the 
functions \code{createStatisticalAdjacencyList} and 
\code{consensusAdjacency}. See \code{?createStatisticalAdjacencyList} and 
\code{?consensusAdjacency} for further details. The function 
\code{createStatisticalAdjacencyList}
includes functionality to caluclate adjacency matrices based on 
LASSO (L1 norm)-regression, random forests, context likelihood of 
relatedness (CLR), the algorithm for the reconstruction of accurate 
cellular networks (ARACNE), Pearson correlation (also partial and 
semipartial), Spearman correlation (also partial and semipartial)
and Constraint-based structure learning (Bayes).
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
createStatisticalAdjacency(x, c("pearson", "spearman"))
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
