% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{.assays_have_identical_colnames_rownames}
\alias{.assays_have_identical_colnames_rownames}
\title{Check if all the assays in the `AdjacencyMatrix` object have
identical colnames and rownames}
\usage{
.assays_have_identical_colnames_rownames(object)
}
\arguments{
\item{object}{`AdjacencyMatrix` object}
}
\value{
`logical` of length 1
}
\description{
The function will check if all the assays in the `AdjacencyMatrix` object
have identical colnames and rownames.
}
\details{
Helper function for validity check of `AdjacencyMatrix` objects.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
