% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{.assays_have_identical_dimnames}
\alias{.assays_have_identical_dimnames}
\title{Check if the assays in the `AdjacencyMatrix` object have
identical dimnames}
\usage{
.assays_have_identical_dimnames(object)
}
\arguments{
\item{object}{`AdjacencyMatrix` object}
}
\value{
`logical` of length 1
}
\description{
The function will check if the assays in the `AdjacencyMatrix` object
have identical dimnames.
}
\details{
Helper function for validity check of `AdjacencyMatrix` objects.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
