% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{exportNet2gml}
\alias{exportNet2gml}
\title{exportNet2gml is a function that exports adjacency matrices to gml using igraph
Needs following attributes:
x: adjacency matrix that needs to be exported
from: originated from wich function, possible values are 
- "structural" Produces a gml file with edge attributes containing mass difference values, data saved as "structural_type.gml"
- "statistical+p" produces a gml file with edge attributes containing correlation values and p-values, saved as "statistical.'model'.gml"
(TO DO: ADD "statistical")
- "combine" produces a gml file with edge attributes containing correlation and p-values for pearson/ spearman correlations, 
saved as "combined.gml"}
\usage{
exportNet2gml(x, from, adjacency_list, ...)
}
\description{
exportNet2gml is a function that exports adjacency matrices to gml using igraph
Needs following attributes:
x: adjacency matrix that needs to be exported
from: originated from wich function, possible values are 
- "structural" Produces a gml file with edge attributes containing mass difference values, data saved as "structural_type.gml"
- "statistical+p" produces a gml file with edge attributes containing correlation values and p-values, saved as "statistical.'model'.gml"
(TO DO: ADD "statistical")
- "combine" produces a gml file with edge attributes containing correlation and p-values for pearson/ spearman correlations, 
saved as "combined.gml"
}
