% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{in_range_which}
\alias{in_range_which}
\title{in_range_which}
\usage{
in_range_which(m_1, m_2, functional_groups)
}
\arguments{
\item{m_1}{numeric, value of the first m/z range}

\item{m_2}{numeric, value of the second m/z range}

\item{functional_groups}{numeric, m/z values of the functional groups}
}
\value{
numeric
}
\description{
The function \code{in_range_which} checks if a m/z feature
(given as a range defined by \code{m_1} and \code{m_2}) is present in a 
list of m/z features given by the \code{functional_groups} argument.
}
\details{
The function \code{in_range_which} is a helper function for 
\code{create_structural_network}. \code{in_range_which} returns the index of
the element in \code{functional_groups} that is in the range defined by 
\code{m_1} and \code{m_2}.
}
\examples{
m_1 <- 180.155
m_2 <- 180.157
functional_groups <- c(164.16, 180.156)
in_range_which(m_1, m_2, functional_groups)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
