% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{inRangeWhich}
\alias{inRangeWhich}
\title{inRangeWhich}
\usage{
inRangeWhich(m_1, m_2, transformation)
}
\arguments{
\item{m_1}{numeric, value of the first m/z range}

\item{m_2}{numeric, value of the second m/z range}

\item{transformation}{numeric, m/z values of the functional groups}
}
\value{
numeric
}
\description{
The function \code{inRangeWhich} checks if a m/z feature
(given as a range defined by \code{m_1} and \code{m_2}) is present in a 
list of m/z features given by the \code{transformation} argument.
}
\details{
The function \code{inRangeWhich} is a helper function for 
\code{createStructuralAdjacency}. \code{inRangeWhich} returns the index of
the element in \code{transformation} that is in the range defined by 
\code{m_1} and \code{m_2}.
}
\examples{
m_1 <- 180.155
m_2 <- 180.157
transformation <- c(164.16, 180.156)
inRangeWhich(m_1, m_2, transformation)
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
