\name{MetNet-package}
\alias{MetNet-package}
\alias{MetNet}
\docType{package}
\title{
\packageTitle{MetNet}
}
\description{
\packageDescription{MetNet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MetNet}
\packageIndices{MetNet}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MetNet}

Maintainer: \packageMaintainer{MetNet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
