% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_test-data.R
\docType{data}
\name{ms2_test}
\alias{ms2_test}
\alias{sps_sub}
\title{Spectra data to test addSpectralSimilarity}
\format{
\code{Spectra}
}
\value{
\code{Spectra}
}
\description{
\code{ms2_test} contains a subset of a \code{Spectra} object. 
It will be used as an example annotation in the vignette to show the 
functionality of the package.
}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
