% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{randomForest}
\alias{randomForest}
\title{Create a statistical network based on random forest}
\usage{
randomForest(x, parallel = FALSE, randomforest_adjust = NULL, ...)
}
\arguments{
\item{x}{matrix, where columns are the samples and the rows are features 
(metabolites), cell entries are intensity values}

\item{parallel}{logical, should computation be parallelized? If 
\code{parallel = TRUE} the \code{mclapply} will be applied if 
\code{parallel = FALSE} the \code{lapply} function will be applied.}

\item{...}{parameters passed to \code{corr.test} and \code{mclapply} (if 
\code{parallel = TRUE})}
}
\value{
matrix, matrix with edges inferred from random forest algorithm 
\code{rfPermute} and \code{rp.importance}
}
\description{
\code{randomForest} infers a statistical network using 
random forest using the \code{rfPermute} function from the 
\code{rfPermute} package. \code{randomForest} extracts the p-values  
by the function \code{rp.importance} and the presence/absence based on the 
significance value ($\alpha$ <= 0.05) of this 
connection to a matrix that is returned.
}
\details{
For use of the parameters used in the \code{rfPermute} function, 
refer to ?rfPermute::rfPermute.
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
\dontrun{randomForest(x)}
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
