% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{statistical}
\alias{statistical}
\title{Create a list of adjacency matrices from statistical methods}
\usage{
statistical(x, model, ...)
}
\arguments{
\item{x}{`matrix` that contains intensity values of 
features/metabolites (rows) per sample (columns).}

\item{model}{`character` vector containing the methods that will be used 
(`"lasso"`, `"randomForest"`, `"clr"`, `"aracne"`, `"pearson"`, 
`"pearson_partial"`, `"pearson_semipartial"`, `"spearman"`, 
`"spearman_partial"`, `"spearman_semipartial"`, `"bayes"`)}

\item{...}{parameters passed to the functions  `lasso`, `randomForest`, 
`clr`, `aracne`, `correlation` and/or `bayes`}
}
\value{
`list` containing the respective adjacency matrices specified by 
`model`
}
\description{
The function `statitical` infers adjacency matrix topologies from 
statistical methods and returns matrices of these networks in a `list`. The 
function includes functionality to calculate adjacency matrices based on 
LASSO (L1 norm)-regression, random forests, context likelihood of 
relatedness (CLR), the algorithm for the reconstruction of accurate 
cellular networks (ARACNE), Pearson correlation (also partial and 
semipartial), Spearman correlation (also partial and semipartial) 
and score-based structure learning (Bayes). The function returns a 
list of adjacency matrices that are defined by `model`.
}
\details{
The function `statistical` includes functionality to calculate adjacency 
matrices based on 
LASSO (L1 norm)-regression, random forests, context likelihood of 
relatedness (CLR), the algorithm for the reconstruction of accurate 
cellular networks (ARACNE), Pearson correlation (also partial and 
semipartial), Spearman correlation (also partial and semipartial)
and Constraint-based structure learning (Bayes). 

`statistical` calls the function
`lasso`, `randomForest`, `clr`, `aracne`, 
`correlation` (for `"pearson"`, `"pearson_partial"`, `"pearson_semipartial"`,
`"spearman"`, `"spearman_partial"`, `"spearman_semipartial"`) and/or `bayes` 
as specified by `model`. It will create adjacency matrices using the 
specified methods and will return a `list` containing the weighted
adjacency matrices. 

Internally `x` will be z-scaled and the z-scaled object 
will be used in `lasso`, `clr` and/or `aracne`.
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
createStatisticalAdjacencyList(x, c("pearson", "spearman"))

}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
