% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{sum_mass}
\alias{sum_mass}
\title{Create a summary from adjacency list containing mass-differences}
\usage{
sum_mass(adjacency_list)
}
\arguments{
\item{x}{is a list of the mass-difference adjacency mateix, that has previously been generated 
by the function `adjacency_list(x, from = "statistical")`or 
`adjacency_list(x, from = "combine")`}
}
\value{
`data.frame` containing the listed values each pair of features from 
adjacency matrix `x` specified by `from`.
}
\description{
The function `sum_mass` creates a summary from the adjacency list.
Individual mass differences are count and summarised over all features. 
The input may be adjacency lists originating from the function `structural`,
or `combine`.
}
\details{
sum_mass summarises the adjacency list containing mass difference values, 
i.e. either adjacency list from structural or combine may be used
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
stat_adj <- statistical(x = x, model = c("pearson", "spearman"))
adj_l <- adjacency_list(x = stat_adj, from = "statistical")
sum_mass(adj_l)

}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
