% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{summary_mz}
\alias{summary_mz}
\title{Create a summary from adjacency list containing mass-differences with possible filter}
\usage{
summary_mz(adjacency_list, filter = F, ...)
}
\arguments{
\item{adjacency_list}{`data.frame`, a list of the mass-difference adjacency mateix, that has previously 
been generated by the function `adjacency_list(x, from = "statistical")`or 
`adjacency_list(x, from = "combine")`}

\item{filter}{`number`/`FALSE`, leave empty or set to `FALSE` if unfiltered data are 
required. Select a `number` as a threshold where mz differences were filtered.
May be usefull at visualization (plotting) for big data.}
}
\value{
`data.frame` containing the numbers of present mz differences and
corresponding name. Also a plot will be displayed.
}
\description{
The function `summary_mz` creates a summary from the adjacency list.
Individual mass differences are count over all features. The input may be 
adjacency lists originating from the function `structural`,
or `combine`. The parameter `filtered` will define if data will be 
filtered above a certain threshold or not.
}
\details{
Summarises the adjacency list containing mass difference values, 
i.e. either adjacency list from structural or combine may be used.
Also the plots will be displayed. 
The default is filter = F, so the unfiltered summary will be returned. 
If filter is set to a `number`, e.g. 1000 only mz differences above 
this threshold will be displayed. 
The function can be applied for adjacency lists from `structural` and `combine`
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
stat_adj <- statistical(x = x, model = c("pearson", "spearman"))
adj_l <- adjacency_list(x = stat_adj, from = "statistical")
summary_mz(adjacency_list = adj_l) 
summary_mz(adjacency_list = adj_l, filter = 100) # filtered


}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
