% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{threeDots_call}
\alias{threeDots_call}
\title{Check if passed arguments match the function's formal arguments and
call the function with the checked arguments}
\usage{
threeDots_call(fun, ...)
}
\arguments{
\item{fun}{function to check for arguments and to call}

\item{...}{arguments to be tested to be passed to fun}
}
\value{
Function call with passed arguments
}
\description{
The function \code{threeDots_call} gets the formal arguments
of a function \code{fun} and checks if the passed arguments \code{...} 
matches the formal arguments. \code{threeDots_call} will remove 
duplicated arguments. \code{threeDots_call} will call the function 
\code{fun} with the filtered arguments and will return the result.
}
\details{
Used internally in \code{lasso}, \code{randomForest}, 
\code{correlation}, \code{bayes}, \code{consensus_network}
}
\examples{
threeDots_call("sd", x = 1:10, y = 1:10)
## in contrast to the above example, the following example will result in an 
## error
\\donotrun{sd(x = 1:10, y = 1:10)}
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
