% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{threeDotsCall}
\alias{threeDotsCall}
\title{Check if passed arguments match the function's formal arguments and
call the function with the checked arguments}
\usage{
threeDotsCall(fun, ...)
}
\arguments{
\item{fun}{`function` to check for arguments and to call}

\item{...}{arguments to be tested to be passed to `fun`}
}
\value{
Returned object given the function call with passed arguments
}
\description{
The function `threeDotsCall` gets the formal arguments
of a function `fun` and checks if the passed arguments `...`
matches the formal arguments. `threeDotsCall` will call the function
`fun` with the filtered arguments and will return the result of the function
call and the given arguments.
}
\details{
Used internally in `lasso`, `randomForest`, `bayes`,
`statistical` and `threshold`.

`threeDotsCall` will not remove duplicated arguments and throw an error.
}
\examples{
MetNet:::threeDotsCall(stats::sd, x = 1:10, y = 1:10)
## in contrast to the above example, the following example will result in an
## error
\dontrun{stats::sd(x = 1:10, y = 1:10)}
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
