% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{threshold_p}
\alias{threshold_p}
\title{Threshold the statistical adjacency matrices (using p-values)}
\usage{
threshold_p(statistical, args, ...)
}
\arguments{
\item{statistical}{`list` containing adjacency matrices that contains p-values}

\item{args}{`list` of arguments, has to contain thresholds for weighted
adjacency matrices depending on the statistical model
(a named list, where names are identical to `model`s in `statistical`)
For example "pearson_p" is the corresponding name for pearson p-values}

\item{...}{parameters passed to the function `consensus` in the
`sna` package}
}
\value{
`matrix`, binary adjacency matrix given the links supported by the `args`
}
\description{
The function `threshold_p` takes as input a list of adjacency matrices
as returned from the function `statistical`. 'threshold_p` will identify the 
strongest link that are lower then a certain p-value threshold
}
\details{
`args` has to contain numeric vector of
length 1 with names equal to `names(statistical)` for each `model`
(`names(statistical)`) and the entry `threshold`, a numerical `vector(1)`
to threshold the consensus
matrix after using the `consensus` function from the `sna` package.


When combining the adjacency matrices the `threshold` value defines if an 
edge is reported or not. The value should be carefully defined by the user.
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[1:10, 3:ncol(x_test)]
x <- as.matrix(x)
model <- c("pearson", "spearman")
l <- statistical(x, model = model, p = TRUE)

args <- list("pearson_p" = 0.05, "spearman" = 0.05, threshold = 1)
threshold_p(statistical = l, type = "threshold", args = args)


}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
