% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{aracne}
\alias{aracne}
\title{Create an adjacency matrix based on algorithm for the reconstruction
of accurate cellular networks}
\usage{
aracne(mi, eps = 0.05, aracne_threshold = 0)
}
\arguments{
\item{mi}{matrix, where columns and the rows are features 
(metabolites), cell entries are mutual information values between the 
features. As input, the mutual information (e.g. raw MI estimates or 
Jackknife bias corrected MI estimates) from the \code{cmi} function of the
\code{mpmi} package can be used.}

\item{eps}{numeric, used to remove the weakest edge of each triple of nodes}

\item{aracne_threshold}{numeric, if the aracne value exceeds the threshold 
(aracne$_{i,j}$ > threshold, where aracne$_{i, j}$ is the aracne value of 
the ith row feature and of the jth column feature), the connection is 
defined as present, if the aracne value is lower than the threshold value 
(aracne$_{i,j}$ <=  threshold) there is no statistical connection reported.}
}
\value{
matrix, matrix with edges inferred from Reconstruction of accurate
cellular networks algorithm \code{aracne}
}
\description{
\code{.information} infers an adjacency matrix using 
the algorithm for the reconstruction of accurate cellular networks 
using the \code{aracne.a} function from the 
\code{parmigene} package. The presence/absence is based on if the 
returned value exceeds a user-defined threshold value. \code{aracne} will 
return the adjacency matrix containing the presence/absence value.
}
\details{
For more details on the \code{aracne.a} function, 
refer to ?parmigene::aracne.a.
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
x_z <- t(apply(x, 1, function(y) (y - mean(y)) / sd(y)))
mi_x_z <- mpmi::cmi(x_z)$bcmi
aracne(mi_x_z, eps = 0.05, aracne_threshold = 0)
}
\author{
Thomas Naake, \email{thomasnaake @googlemail.com}
}
