% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{clr}
\alias{clr}
\title{Create an adjacency matrix based on context likelihood or 
relatedness network}
\usage{
clr(mi, clr_threshold=0)
}
\arguments{
\item{mi}{matrix, where columns and the rows are features 
(metabolites), cell entries are mutual information values between the 
features. As input, the mutual information (e.g. raw MI estimates or 
Jackknife bias corrected MI estimates) from the \code{cmi} function of the
\code{mpmi} package can be used.}

\item{clr_threshold}{numeric, if the clr value exceeds the threshold 
(clr$_{i,j}$ > threshold, where clr$_{i, j}$ is the clr value of the ith row 
feature and of the jth column feature), the connection is defined as 
present, if the clr value is lower than the threshold value 
(clr$_{i,j}$ \eqn{\leq} threshold)
there is no statistical connection reported.}
}
\value{
matrix, matrix with edges inferred from Context Likelihood or 
Relatedness Network algorithm \code{clr}
}
\description{
\code{clr} infers an adjacency matrix using 
context likelihood/relatedness network using the \code{clr} function from 
the \code{parmigene} package. The presence/absence is based on if the 
returned value exceeds a user-defined threshold value. \code{clr} will 
return the adjacency matrix containing the presence/absence value.
}
\details{
For more details on the \code{clr} function, 
refer to ?parmigene::clr.
}
\examples{
data("x_test", package="MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
x_z <- t(apply(x, 1, function(y) (y - mean(y)) / sd(y)))
mi_x_z <- mpmi::cmi(x_z)$bcmi
clr(mi_x_z, clr_threshold=0)
}
\author{
Thomas Naake, \email{thomasnaake @googlemail.com}
}
