% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{correlation}
\alias{correlation}
\title{Create a statistical network based on correlation}
\usage{
correlation(x, adjust_correlation = "none", threshold_correlation = 0.05, ...)
}
\arguments{
\item{x}{matrix, where columns are the samples and the rows are features 
(metabolites), cell entries are intensity values}

\item{adjust_correlation}{character}

\item{threshold_correlation}{numeric, significance level $\alpha$ 
(default: 0.05), if the (adjusted) p-values exceed this value, there 
is no statistical connection between features}

\item{...}{parameters passed to \code{corr.test} (argument \code{adjust} will be ignored)}
}
\value{
matrix, matrix with edges inferred from correlation algorithm 
\code{corr.test}
}
\description{
\code{correlation} infers a statistical network using 
correlation using the \code{corr.test} function from the 
\code{psych} package. \code{correlation} extracts the reported 
p-values from the function \code{corr.test} that can be adjusted for 
multiple testing (\code{adjust_correlation} parameter).
}
\details{
For use of the parameters used in the \code{corr.test} function, 
refer to ?psych::corr.test.
}
\examples{
data("x_test", package = "MetNet")
x <- x_test[, 3:dim(x_test)[2]]
x <- as.matrix(x)
correlation(x, adjust_correlation = "bonferroni", type = "pearson")
}
\author{
Thomas Naake, \email{thomasnaake @googlemail.com}
}
