% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{sum_mass}
\alias{sum_mass}
\title{sum_mass summarises the adjacency list containing mass difference values, 
i.e. either adjacency list from structural or combine may be used}
\usage{
sum_mass(adjacency_list)
}
\description{
sum_mass summarises the adjacency list containing mass difference values, 
i.e. either adjacency list from structural or combine may be used
}
